fix four pre-existing clippy lints in ramus-core test code
ci runs clippy with -D warnings, which promoted these to errors:
- unused AtomicUsize import in connection.rs tests
- redundant `as i32` cast on i32 in db.rs test fixture
- two `vec\![...]` -> `[...]` in media_keys.rs and mpv.rs tests

oops lol, locally i only ran clippy on ramus-tauri and missed
these. all four are test-code only, no behavior change.

+++ ramus-core/src/cache/db.rs
@@ -654,7 +654,7 @@ mod tests {
             album_items.push(AlbumUpsertRow {
                 title: format!("Album {}", i),
                 artist_id: first_artist_id,
-                year: Some(2000 + (i % 20) as i32),
+                year: Some(2000 + (i % 20)),
                 source_id: format!("al{}", i),
                 art_url: None,
                 updated_at: Some(1000),
